pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-orchestrator"
version = "0.1.0"
description = "Generic YAML-Driven ReAct Agent Orchestration Engine"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "jsonschema>=4.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "langfuse>=3.0.0",
    "mcp>=1.27.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "jsonpath-ng>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["engine*"]
exclude = ["configs*", "output*", "transcripts*"]
